+++ xrpl/utils/currency_converters.go
@@ -1,4 +1,4 @@
-package helpers
+package utils
 
 import (
 	"encoding/hex"
