+++ xrpl/utils/currency_converters_test.go
@@ -1,4 +1,4 @@
-package helpers
+package utils
 
 import (
 	"testing"
